Added a new fast but sub-optimal pathfinding
It performs a flood fill from the target back to the source. It reuses
the same flood fill map for all queries to the same target, as long as
the produced next step in the path is legal. This might result in
suboptimal paths when a unit blocking a shorter path moves out of the
way, but the cache still returns the previous one. this pathfinding is
really usefull on bigger maps.